Replace "him or her" in Code of Conduct
There doesn't seem to be any reason to use "him or her" as language in this code of conduct.  "Them" works just as well in this sentence and doesn't suggest that there are only "hims" and "hers" offending people.

I understand this a totally trivial change, but I would totally appreciate it if the Code of Conduct wasn't upholding a "false" binary gender system.  Thanks!